Rework the System which makes items unclickable after buy
The first approach was to Destroy the BuyableObject component of the
clone after placing it. This couldnt be tested in EditMode tests and
DestroyImmediate was no alternative due to doc recommendations to
not use it. Instead, the BuyableObject.isAlreadyBought bool was
added. Setting it to true via the public method BuyableObject.Buy()
will prevent click interaction. Counter.PlaceOnCounter calls it on
the clone after buy.